+++ pages/options.js
@@ -5,6 +5,7 @@
     useProfile: false,
     steamId: "",
     showTrailer: true,
+    showProfileCards: true,
     language: "auto",
   };
   document.documentElement.lang = chrome.i18n.getUILanguage().replace("_", "-");